Add files via upload
The only change needed to get this working was the Spread Sig. I legit have not slept for maybe like 50ish hours, I lost track of a lot and had to restart what Id done a couple times because I got lost at where I was at with trying to learn assembly and how to navigate memory from Cheat Engine. Im pretty sure this version is still only changed colors, however I think there is also a button for grappling in the menu. It does absolutely nothing, I just cant be bothered at the moment to mess with the code more when it doesnt cause any issues.