Add files via upload
This version only cleans the menu up a bit and adds a few things, the biggest is instant infiltrations thanks to Athrun, but I dont know how safe it is to use it as it directly sends requests to the server. The others are just customizable things, like picking between Square or Circle shapes in ESP, Shape Size slider as well as a Shape Distance Scaler with the Shapes sizes capped out at being too big or small. I spent hours trying to get a way to customize the colors of the GUI while in game from the GUI Menu, but no matter what I did it wouldn't apply changes made in the GUI. I may still look for a way to do it, but if you want to change the colors just open the main DLL file and search for ColorPink or ColorDarkRed, they are used a couple of times but one will have all the colors listed around it and you can set them to what youd like, just look up RGBA Float values for colors.